+++ src/Controller/OrderController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Repository\OrdersRepository;
 Use App\Repository\ProductRepository;
 use App\Repository\CustomerRepository;
+use ErrorException;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -38,28 +39,40 @@ class OrderController extends AbstractController
     {
         $data = json_decode($request->getContent(), true);
        
-        $customerId = $data["customer_id"];
-        $productId = $data["product_id"];
-        $price = $data["price"];
-        $customer = $this->customerRepository->findOneBy(['id'=>$customerId]);
-        $product = $this->productRepository->findOneBy(['id'=>$productId]);
-       
-        
-        if (empty($customerId) || empty($productId) || empty($price) ){
-            throw new NotFoundHttpException("Expecting mandatory parameters:");
-        }
-        
-        
-        $id = $this->orderRepository->saveOrder($customer, $product);
-        return new JsonResponse(['message' => "order created with id number ".$id], Response::HTTP_CREATED);
+        try {
+
+            $customerId = $data["customer_id"];
+            $productId = $data["product_id"];
+            $price = $data["price"];
+            $customer = $this->customerRepository->findOneBy(['id'=>$customerId]);
+            $product = $this->productRepository->findOneBy(['id'=>$productId]);
+
+            $id = $this->orderRepository->saveOrder($customer, $product);
+
+            return new JsonResponse(['message' => "order created with id number ".$id], Response::HTTP_CREATED);
+
+        } catch(ErrorException $e) {
+            
+            return new JsonResponse(["Error message" => "Enter valid arguments"]);
+
+        } catch(\Doctrine\DBAL\Exception\NotNullConstraintViolationException $e) {
+
+            return new JsonResponse(["Error message" => "Enter valid arguments"]);
+        }      
     }
 
     /**
      * @Route("/{id}", name="get_one_customer", methods={"GET"})
      */
     public function getOneOrder($id): JsonResponse
-    {
+    {   
         $order = $this->orderRepository->findOneBy(['id' => $id]);
+
+        if (is_null($order)) {
+
+            return new JsonResponse(["Error message" => "Order not found"]);
+        }
+
         $customerName = $order->getCustomerId()->getFirstName();
         $productName = $order->getProductId()->getName();
         $data = [
@@ -67,7 +80,6 @@ class OrderController extends AbstractController
             'customer' => $customerName,
             'product' => $productName,
             'price' => $order->getProductId()->getPrice(),
-            
         ];
 
         return new JsonResponse(['order' => $data], Response::HTTP_OK);
@@ -79,6 +91,13 @@ class OrderController extends AbstractController
     public function deleteOrder($id) : JsonResponse
     {
         $order = $this->orderRepository->findOneBy(['id'=>$id]);
+
+        if (is_null($order)) {
+
+            return new JsonResponse(["Error message" => "Order not found"]);
+
+        }
+
         $this->orderRepository->removeOrder($order);
 
         return new JsonResponse(["message"=>"order deleted with id number ".$id]);
@@ -92,9 +111,18 @@ class OrderController extends AbstractController
     public function updateOrder($id, Request $request) :JsonResponse
     {
         $order = $this->orderRepository->findOneBy(["id"=>$id]);
+
+        if (is_null($order)) {
+
+            return new JsonResponse(["Error message" => "Order not found"]);
+
+        }
+
         $data = json_decode($request->getContent(), true);
 
-
+        if (is_null($data)){
+            return new JsonResponse(["Error message" => "Enter data to update"]);
+        }
 
         if (!empty($data['customer_id'])){
             $data["customer_id"]=$this->customerRepository->findOneBy(["id"=>$data["customer_id"]]);
@@ -108,8 +136,7 @@ class OrderController extends AbstractController
         {
             $data["price"] = $this->productRepository->findOneBy(["id"=>$data["product_id"]]);
         }
-
-        
+ 
         $this->orderRepository->updateOrder($order, $data);
         return new JsonResponse(["message"=> "Order updated with id number ".$id]);
     }
